Support for window space using full UnsignedInteger32 value (instead of Integer.MAX_VALUE)

+++ maverick-synergy-common/src/main/java/com/sshtools/synergy/ssh/SocketForwardingChannel.java
@@ -80,7 +80,7 @@ public abstract class SocketForwardingChannel<T extends SshContext> extends Forw
 				con.getContext().getPolicy(ForwardingPolicy.class).getForwardingMaxWindowSize(),
 				con.getContext().getPolicy(ForwardingPolicy.class).getForwardingMaxWindowSize(), 
 				con.getContext().getPolicy(ForwardingPolicy.class).getForwardingMinWindowSize());
-		toChannel = new ForwardingDataWindow(con.getContext().getPolicy(ForwardingPolicy.class).getForwardingMaxWindowSize());
+		toChannel = new ForwardingDataWindow(con.getContext().getPolicy(ForwardingPolicy.class).getForwardingMaxWindowSize().intValue());
 	}
 
 	protected CachingDataWindow createCache(int maximumWindowSpace) {
